Refactor carousel to allow multiple buttons
