+++ src/util/styles.ts
@@ -1,6 +1,8 @@
 import styled from "styled-components";
 import COLORS from "./colors";
 
+export type ButtonStyle = "danger" | "default";
+
 export const BaseButton = styled.button({
   border: `solid 3px ${COLORS.foregroundPrimary}`,
   color: `${COLORS.foregroundPrimary}`,
@@ -15,7 +17,7 @@ export const BaseButton = styled.button({
   [`&[disabled], &[aria-disabled="true"]`]: { opacity: "10%", filter: "grayscale(100%)" },
 });
 
-export const ActionButton = styled.button<{ buttonStyle?: "danger" | "default" }>(({ buttonStyle = "default" }) => {
+export const ActionButton = styled.button<{ buttonStyle?: ButtonStyle }>(({ buttonStyle = "default" }) => {
   const rgbMap: { [k in typeof buttonStyle]: string } = { default: "0, 150, 0", danger: "196, 0, 0" };
   return {
     padding: "12px",
@@ -23,7 +25,7 @@ export const ActionButton = styled.button<{ buttonStyle?: "danger" | "default" }
     borderRadius: "5px",
     fontSize: "18px",
     fontWeight: "bold",
-    color: `${COLORS.foregroundBase}`,
+    color: COLORS.foregroundBase,
     backgroundColor: `rgb(${rgbMap[buttonStyle]})`,
     "&:active": { filter: "brightness(0.7)" },
     [`&[disabled], &[aria-disabled="true"]`]: { opacity: "10%", filter: "grayscale(100%)" },
